+++ backend/tests/Functional/Controller/User/ServiceControllerTest_TestService.php
@@ -14,7 +14,7 @@ class ServiceControllerTest_TestService implements ServiceInterface
 {
     public static $lastGroups;
 
-    public function __construct(LoggerInterface $logger)
+    public function __construct(LoggerInterface $logger, string $configurationData)
     {
     }